updating using knn function
Doubt in the last cell. I was trying to do it using caret package, but it is giving an error that trainControl function is not found. I checked everywhere on the net, I guess the problem is with the R version in Jupyter Notebook, but I am not sure. Please help here and the best AUC that I am getting in .73, even from knn function. Is there any way we can improve this?